Develop

+++ app/src/main/java/com/luacevedo/heartbaymax/model/rules/conditions/GreaterThanCondition.java
@@ -2,21 +2,21 @@ package com.luacevedo.heartbaymax.model.rules.conditions;
 
 import com.luacevedo.heartbaymax.model.patient.PatientAttribute;
 
-public class GreaterThanCondition extends BaseCondition<Integer> {
+public class GreaterThanCondition extends BaseCondition<Double> {
 
-    private int min;
+    private Double min;
 
-    public GreaterThanCondition(String attributeRoot, int min) {
+    public GreaterThanCondition(String attributeRoot, Double min) {
         this.attributeRoot = attributeRoot;
         this.min = min;
     }
 
-    public void setMin(Integer min) {
+    public void setMin(Double min) {
         this.min = min;
     }
 
     @Override
-    public boolean validate(PatientAttribute<Integer> attribute) {
+    public boolean validate(PatientAttribute<Double> attribute) {
         return attribute.getValue() > min;
     }
 }
